# internal/protocols/tuist_cache/openapi/oas_router_gen.go
// Code generated by ogen, DO NOT EDIT.

package openapi

import (
	"net/http"
	"net/url"
	"strings"

	"github.com/ogen-go/ogen/uri"
)

func (s *Server) cutPrefix(path string) (string, bool) {
	prefix := s.cfg.Prefix
	if prefix == "" {
		return path, true
	}
	if !strings.HasPrefix(path, prefix) {
		// Prefix doesn't match.
		return "", false
	}
	// Cut prefix from the path.
	return strings.TrimPrefix(path, prefix), true
}

// ServeHTTP serves http request as defined by OpenAPI v3 specification,
// calling handler that matches the path or returning not found error.
func (s *Server) ServeHTTP(w http.ResponseWriter, r *http.Request) {
	elem := r.URL.Path
	elemIsEscaped := false
	if rawPath := r.URL.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
			elemIsEscaped = strings.ContainsRune(elem, '%')
		}
	}

	elem, ok := s.cutPrefix(elem)
	if !ok || len(elem) == 0 {
		s.notFound(w, r)
		return
	}
	args := [1]string{}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/api/cache/"

			if l := len("/api/cache/"); len(elem) >= l && elem[0:l] == "/api/cache/" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				break
			}
			switch elem[0] {
			case 'c': // Prefix: "c"

				if l := len("c"); len(elem) >= l && elem[0:l] == "c" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'a': // Prefix: "as/"

					if l := len("as/"); len(elem) >= l && elem[0:l] == "as/" {
						elem = elem[l:]
					} else {
						break
					}

					// Param: "id"
					// Leaf parameter, slashes are prohibited
					idx := strings.IndexByte(elem, '/')
					if idx >= 0 {
						break
					}
					args[0] = elem
					elem = ""

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "GET":
							s.handleDownloadCASArtifactRequest([1]string{
								args[0],
							}, elemIsEscaped, w, r)
						case "POST":
							s.handleSaveCASArtifactRequest([1]string{
								args[0],
							}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, "GET,POST")
						}

						return
					}

				case 'l': // Prefix: "lean"

					if l := len("lean"); len(elem) >= l && elem[0:l] == "lean" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "DELETE":
							s.handleCleanProjectCacheRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, "DELETE")
						}

						return
					}

				}

			case 'k': // Prefix: "keyvalue"

				if l := len("keyvalue"); len(elem) >= l && elem[0:l] == "keyvalue" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch r.Method {
					case "PUT":
						s.handlePutKeyValueRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, "PUT")
					}

					return
				}
				switch elem[0] {
				case '/': // Prefix: "/"

					if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
						elem = elem[l:]
					} else {
						break
					}

					// Param: "cas_id"
					// Leaf parameter, slashes are prohibited
					idx := strings.IndexByte(elem, '/')
					if idx >= 0 {
						break
					}
					args[0] = elem
					elem = ""

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "GET":
							s.handleGetKeyValueRequest([1]string{
								args[0],
							}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, "GET")
						}

						return
					}

				}

			case 'm': // Prefix: "module/"

				if l := len("module/"); len(elem) >= l && elem[0:l] == "module/" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'c': // Prefix: "complete"
					origElem := elem
					if l := len("complete"); len(elem) >= l && elem[0:l] == "complete" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "POST":
							s.handleCompleteModuleCacheMultipartUploadRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, "POST")
						}

						return
					}

					elem = origElem
				case 'p': // Prefix: "part"
					origElem := elem
					if l := len("part"); len(elem) >= l && elem[0:l] == "part" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "POST":
							s.handleUploadModuleCachePartRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, "POST")
						}

						return
					}

					elem = origElem
				case 's': // Prefix: "start"
					origElem := elem
					if l := len("start"); len(elem) >= l && elem[0:l] == "start" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "POST":
							s.handleStartModuleCacheMultipartUploadRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, "POST")
						}

						return
					}

					elem = origElem
				}
				// Param: "id"
				// Leaf parameter, slashes are prohibited
				idx := strings.IndexByte(elem, '/')
				if idx >= 0 {
					break
				}
				args[0] = elem
				elem = ""

				if len(elem) == 0 {
					// Leaf node.
					switch r.Method {
					case "GET":
						s.handleDownloadModuleCacheArtifactRequest([1]string{
							args[0],
						}, elemIsEscaped, w, r)
					case "HEAD":
						s.handleModuleCacheArtifactExistsRequest([1]string{
							args[0],
						}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, "GET,HEAD")
					}

					return
				}

			}

		}
	}
	s.notFound(w, r)
}

// Route is route object.
type Route struct {
	name           string
	summary        string
	operationID    string
	operationGroup string
	pathPattern    string
	count          int
	args           [1]string
}

// Name returns ogen operation name.
//
// It is guaranteed to be unique and not empty.
func (r Route) Name() string {
	return r.name
}

// Summary returns OpenAPI summary.
func (r Route) Summary() string {
	return r.summary
}

// OperationID returns OpenAPI operationId.
func (r Route) OperationID() string {
	return r.operationID
}

// OperationGroup returns the x-ogen-operation-group value.
func (r Route) OperationGroup() string {
	return r.operationGroup
}

// PathPattern returns OpenAPI path.
func (r Route) PathPattern() string {
	return r.pathPattern
}

// Args returns parsed arguments.
func (r Route) Args() []string {
	return r.args[:r.count]
}

// FindRoute finds Route for given method and path.
//
// Note: this method does not unescape path or handle reserved characters in path properly. Use FindPath instead.
func (s *Server) FindRoute(method, path string) (Route, bool) {
	return s.FindPath(method, &url.URL{Path: path})
}

// FindPath finds Route for given method and URL.
func (s *Server) FindPath(method string, u *url.URL) (r Route, _ bool) {
	var (
		elem = u.Path
		args = r.args
	)
	if rawPath := u.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
		}
		defer func() {
			for i, arg := range r.args[:r.count] {
				if unescaped, err := url.PathUnescape(arg); err == nil {
					r.args[i] = unescaped
				}
			}
		}()
	}

	elem, ok := s.cutPrefix(elem)
	if !ok {
		return r, false
	}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/api/cache/"

			if l := len("/api/cache/"); len(elem) >= l && elem[0:l] == "/api/cache/" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				break
			}
			switch elem[0] {
			case 'c': // Prefix: "c"

				if l := len("c"); len(elem) >= l && elem[0:l] == "c" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'a': // Prefix: "as/"

					if l := len("as/"); len(elem) >= l && elem[0:l] == "as/" {
						elem = elem[l:]
					} else {
						break
					}

					// Param: "id"
					// Leaf parameter, slashes are prohibited
					idx := strings.IndexByte(elem, '/')
					if idx >= 0 {
						break
					}
					args[0] = elem
					elem = ""

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "GET":
							r.name = DownloadCASArtifactOperation
							r.summary = "Download a CAS artifact"
							r.operationID = "downloadCASArtifact"
							r.operationGroup = ""
							r.pathPattern = "/api/cache/cas/{id}"
							r.args = args
							r.count = 1
							return r, true
						case "POST":
							r.name = SaveCASArtifactOperation
							r.summary = "Save a CAS artifact"
							r.operationID = "saveCASArtifact"
							r.operationGroup = ""
							r.pathPattern = "/api/cache/cas/{id}"
							r.args = args
							r.count = 1
							return r, true
						default:
							return
						}
					}

				case 'l': // Prefix: "lean"

					if l := len("lean"); len(elem) >= l && elem[0:l] == "lean" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "DELETE":
							r.name = CleanProjectCacheOperation
							r.summary = "Clean all cache artifacts for a project"
							r.operationID = "cleanProjectCache"
							r.operationGroup = ""
							r.pathPattern = "/api/cache/clean"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}

				}

			case 'k': // Prefix: "keyvalue"

				if l := len("keyvalue"); len(elem) >= l && elem[0:l] == "keyvalue" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch method {
					case "PUT":
						r.name = PutKeyValueOperation
						r.summary = "Put a key-value entry"
						r.operationID = "putKeyValue"
						r.operationGroup = ""
						r.pathPattern = "/api/cache/keyvalue"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}
				switch elem[0] {
				case '/': // Prefix: "/"

					if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
						elem = elem[l:]
					} else {
						break
					}

					// Param: "cas_id"
					// Leaf parameter, slashes are prohibited
					idx := strings.IndexByte(elem, '/')
					if idx >= 0 {
						break
					}
					args[0] = elem
					elem = ""

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "GET":
							r.name = GetKeyValueOperation
							r.summary = "Get a key-value entry"
							r.operationID = "getKeyValue"
							r.operationGroup = ""
							r.pathPattern = "/api/cache/keyvalue/{cas_id}"
							r.args = args
							r.count = 1
							return r, true
						default:
							return
						}
					}

				}

			case 'm': // Prefix: "module/"

				if l := len("module/"); len(elem) >= l && elem[0:l] == "module/" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'c': // Prefix: "complete"
					origElem := elem
					if l := len("complete"); len(elem) >= l && elem[0:l] == "complete" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "POST":
							r.name = CompleteModuleCacheMultipartUploadOperation
							r.summary = "Complete a multipart module cache upload"
							r.operationID = "completeModuleCacheMultipartUpload"
							r.operationGroup = ""
							r.pathPattern = "/api/cache/module/complete"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}

					elem = origElem
				case 'p': // Prefix: "part"
					origElem := elem
					if l := len("part"); len(elem) >= l && elem[0:l] == "part" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "POST":
							r.name = UploadModuleCachePartOperation
							r.summary = "Upload a part of a multipart module cache upload"
							r.operationID = "uploadModuleCachePart"
							r.operationGroup = ""
							r.pathPattern = "/api/cache/module/part"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}

					elem = origElem
				case 's': // Prefix: "start"
					origElem := elem
					if l := len("start"); len(elem) >= l && elem[0:l] == "start" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "POST":
							r.name = StartModuleCacheMultipartUploadOperation
							r.summary = "Start a multipart module cache upload"
							r.operationID = "startModuleCacheMultipartUpload"
							r.operationGroup = ""
							r.pathPattern = "/api/cache/module/start"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}

					elem = origElem
				}
				// Param: "id"
				// Leaf parameter, slashes are prohibited
				idx := strings.IndexByte(elem, '/')
				if idx >= 0 {
					break
				}
				args[0] = elem
				elem = ""

				if len(elem) == 0 {
					// Leaf node.
					switch method {
					case "GET":
						r.name = DownloadModuleCacheArtifactOperation
						r.summary = "Download a module cache artifact"
						r.operationID = "downloadModuleCacheArtifact"
						r.operationGroup = ""
						r.pathPattern = "/api/cache/module/{id}"
						r.args = args
						r.count = 1
						return r, true
					case "HEAD":
						r.name = ModuleCacheArtifactExistsOperation
						r.summary = "Check if a module cache artifact exists"
						r.operationID = "moduleCacheArtifactExists"
						r.operationGroup = ""
						r.pathPattern = "/api/cache/module/{id}"
						r.args = args
						r.count = 1
						return r, true
					default:
						return
					}
				}

			}

		}
	}
	return r, false
}